pyproject.toml
[project]
name = "schedule-builder-backend"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "authlib>=1.6.0",
    "colorlog>=6.9.0",
    "fastapi>=0.115.12",
    "numpy>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "fastapi-swagger>=0.2.19",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.5",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
]

[tool.ruff]
line-length = 120
lint.ignore = [
    # wildcard imports are fine
    "F403",
    "F405",
    "PLR",
]
lint.extend-select = ["I", "UP", "PL"]
lint.extend-ignore = ["PLC0415"]
target-version = "py312"
